Fixes failing build and temporarily disables build
A lot of the components shown in the design system belonged to the old frontend. This no longer exists.
I had to copy the components to templates/design-system/components to stop the build failing.
We need to go through each of the these and decide if it should be in digital-land-frontend.
If so we should build it up properly in that repo.
We should remove from here.
And if we no longer think it is a reusable component we should remove the examples and doc pages.